feat(B-490): add user_stats endpoint to user_profile router
