+++ backend/app/main.py
@@ -17,7 +17,6 @@ from app.routers import (
     session_route,
     session_turn_route,
     user_profile_route,
-    user_profile_stats_route,
 )
 
 logging.basicConfig(
@@ -42,7 +41,6 @@ app.include_router(conversation_scenario_route.router)
 app.include_router(session_route.router)
 app.include_router(session_turn_route.router)
 app.include_router(user_profile_route.router)
-app.include_router(user_profile_stats_route.router)
 app.include_router(app_config_route.router)
 app.include_router(admin_dashboard_stats_route.router)
 app.include_router(review_route.router)
